GoToScene and WalkPath and Explore with server
implement C/S part of GoToScene and WalkPath and Explore   and refactor
some part of server and client

+++ ResourceEmperorServer/REStructure/PathwayInfo.cs
@@ -4,6 +4,6 @@
     {
         public Scene Scene;
         public int Distance;
-        public int DiscoveredProbability;
+        public double DiscoveredProbability;
     }
 }
